backend/requirements.txt
Django==2.2.3
certifi==2019.6.16
chardet==3.0.4
Click==7.0
djangorestframework==3.10.3
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pytz==2019.1
requests==2.22.0
sqlparse==0.2.4
urllib3==1.25.3
Werkzeug==0.16.0
django-extensions==2.2.5
django-rest-knox==4.1.0
djongo==1.2.36
drf-yasg==1.17.0
django-cors-headers==3.2.0
yara-python==3.11.0